requirements.txt
frida
python-telegram-bot
requests